requirements.txt
langchain==0.3.4
python-dotenv==1.0.1
langchain-openai==0.2.3
langgraph==0.2.39
langchain-community==0.3.3
tavily-python==0.5.0
wikipedia==1.4.0
youtube-search==2.1.2
asknews==0.7.50
chainlit==1.3.0